feat: split completePool into 3 different functions
did this to setWinners and theirs stats in batches preventing loop failures due to gas or block llimit.
use PoolsManager as the feeCollector in finalizePoolCompletion directly.
get rid of the IPoolCompletionHandler and call processPoolCompletion from manager.
use the pool multiplier and fees percent from seeds in computing winners counts and pool fees.